backend/requirements.txt
alembic==1.6.5
fastapi==0.67.0
FastAPI-SQLAlchemy==0.2.1
pydantic==1.8.2
psycopg2-binary==2.9.1
uvicorn==0.15.0
fastapi-auth0==0.2.1
PyJWT==2.3.0
Faker==8.11.0
pytest==6.2.4
pytest-asyncio==0.15.1
databases==0.4.3
httpx==0.18.2
asgi-lifespan==1.0.1
openpyxl==3.0.5
xlrd==1.2.0
flake8==4.0.1
xlwt==1.3.0
pyhumps==3.0.2
google-cloud-storage==1.42.3
aiofiles==0.7.0
alembic-utils==0.7.3
python-multipart==0.0.5
XlsxWriter==3.0.2
fastapi-utils==0.2.1
mailjet-rest==1.3.4
Jinja2==3.0.3
bs4==0.0.1
lxml==4.6.4
jsmin==3.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
numpy==1.22.1
pandas==1.4.0

# code coverage
coverage==6.2
coveralls==3.3.1
inquirer==2.9.2
gevent==22.8.0
